Parse data and parameters of OPUS binary file of a Bruker spectrometer

This function can be used to read and parse OPUS files, to make it usable for other processing steps.

parse_opus(raw, data_only)

Arguments

raw

data source name; can be a file path to an OPUS file or directly a raw vector

data_only

TRUE or FALSE if only the ab or refl data should be returned

Value

list containing the different data chunks of an OPUS file, possibly:

refl_data_param:

refl:

quant_report_refl:

sc_sample_data_param:

sc_sample:

sc_ref_data_param:

sc_ref:

optics:

optics_ref:

acquisition_ref:

fourier_transformation_ref:

fourier_transformation:

sample:

acquisition:

instrument_ref:

instrument:

lab_and_process_param_1:

lab_and_process_param_2:

info_block:

history:

See also

Examples

library(opusreader2)

dsn <- system.file(
  "extdata/test_data/BF_lo_01_soil_cal.1",
  package = "opusreader2"
)

raw <- read_opus_raw(dsn)

opus_list <- parse_opus(raw, data_only = FALSE)